Move methods on top-level Capybara module from dsl.rb to capybara.rb
Do not require "capybara/dsl" in capybara.rb (reverting 1192ea34),
because it crashes from load-ordering issues. However, we now autoload
Capybara::DSL.
